requirements.txt
bs4
fat-forensics
flax
gensim
ipykernel
jax
nltk
numpy
optax
pandas
pyarrow
seaborn
scikit-learn
scipy
setfit
tensorflow
tf-keras
torch
tqdm
